# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Binance Futures Paper Trading Research Platform API
servers:
  - url: /api
    description: Base API path
security:
  - bearerAuth: []
...
paths:
...
  /signals:
    get:
      operationId: getSignals
      tags: [signals]
      summary: Get all signals
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, active, expired, traded]
        - name: direction
          in: query
          schema:
            type: string
            enum: [LONG, SHORT]
        - name: source
          in: query
          schema:
            $ref: "#/components/schemas/SignalSource"
        - name: scannerType
          in: query
          schema:
            $ref: "#/components/schemas/ScannerType"
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Signals list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Signal"
  /signals/{id}:
    get:
      operationId: getSignal
      tags: [signals]
      summary: Get signal by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Signal detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Signal"
        "404":
          description: Not found
...
  /analytics/daily:
    get:
      operationId: getDailyPerformance
      tags: [analytics]
      summary: Get daily performance history
      parameters:
        - name: days
          in: query
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: Daily performance
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DailyPerformance"
...
  /analytics/pnl-curve:
    get:
      operationId: getPnlCurve
      tags: [analytics]
      summary: Get cumulative PnL curve data
      parameters:
        - name: days
          in: query
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: PnL curve
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PnlPoint"
  /analytics/direction-performance:
    get:
      operationId: getDirectionPerformance
      tags: [analytics]
      summary: Get closed trade performance by direction
      parameters:
        - name: days
          in: query
          schema:
            type: integer
            default: 90
            minimum: 1
            maximum: 3650
      responses:
        "200":
          description: Direction performance summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DirectionPerformanceSummary"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SignalSource:
      type: string
      enum: [TECHNICAL, SMC]
    ScannerType:
      type: string
      enum: [TECHNICAL_SCANNER, SMC_SCANNER]
...
    DiagnosticDetails:
      type: object
      additionalProperties: true
...
    Signal:
      type: object
      required: [id, symbol, direction, score, grade, entryPrice, stopLoss, tp1, tp2, tp3, status, reason, createdAt]
      properties:
        id:
          type: integer
        symbol:
          type: string
        direction:
          type: string
          enum: [LONG, SHORT]
        source:
          $ref: "#/components/schemas/SignalSource"
        scannerType:
          $ref: "#/components/schemas/ScannerType"
        strategyType:
          type: string
        strategyLabel:
          type: ["string", "null"]
        badge:
          type: ["string", "null"]
        smcScore:
          type: ["number", "null"]
        smcDetails:
          oneOf:
            - $ref: "#/components/schemas/DiagnosticDetails"
            - type: "null"
        score:
          type: number
        grade:
          type: string
          enum: [A+, A]
        entryPrice:
          type: number
        stopLoss:
          type: number
        tp1:
          type: number
        tp2:
          type: number
        tp3:
          type: number
        rrRatio:
          type: ["number", "null"]
        status:
          type: string
          enum: [pending, active, expired, traded]
        reason:
          type: string
        slReason:
          type: ["string", "null"]
        trendScore:
          type: ["number", "null"]
        structureScore:
          type: ["number", "null"]
        volumeScore:
          type: ["number", "null"]
        breakoutScore:
          type: ["number", "null"]
        retestScore:
          type: ["number", "null"]
        createdAt:
          type: string
        expiresAt:
          type: ["string", "null"]
...
    DirectionTradeSummary:
      type: object
      required: [closedTradeCount, winRate, averagePnl, averageScore, averageDurationMinutes]
      properties:
        closedTradeCount:
          type: integer
        winRate:
          type: number
        averagePnl:
          type: number
        averageScore:
          type: number
        averageDurationMinutes:
          type: number
    SymbolPerformanceSummary:
      type: object
      required: [symbol, direction, totalPnl, averagePnl, averageScore, averageDurationMinutes, winRate, count]
      properties:
        symbol:
          type: string
        direction:
          type: string
          enum: [LONG, SHORT]
        totalPnl:
          type: number
        averagePnl:
          type: number
        averageScore:
          type: number
        averageDurationMinutes:
          type: number
        winRate:
          type: number
        count:
          type: integer
    DirectionPerformanceSummary:
      type: object
      required: [days, from, closedLongTradeCount, closedShortTradeCount, longWinRate, shortWinRate, averageLongPnl, averageShortPnl, averageLongScore, averageShortScore, averageLongDurationMinutes, averageShortDurationMinutes, bestSymbols, worstSymbols, byDirection]
      properties:
        days:
          type: integer
        from:
          type: string
        closedLongTradeCount:
          type: integer
        closedShortTradeCount:
          type: integer
        longWinRate:
          type: number
        shortWinRate:
          type: number
        averageLongPnl:
          type: number
        averageShortPnl:
          type: number
        averageLongScore:
          type: number
        averageShortScore:
          type: number
        averageLongDurationMinutes:
          type: number
        averageShortDurationMinutes:
          type: number
        bestSymbols:
          type: array
          items:
            $ref: "#/components/schemas/SymbolPerformanceSummary"
        worstSymbols:
          type: array
          items:
            $ref: "#/components/schemas/SymbolPerformanceSummary"
        byDirection:
          type: object
          required: [LONG, SHORT]
          properties:
            LONG:
              $ref: "#/components/schemas/DirectionTradeSummary"
            SHORT:
              $ref: "#/components/schemas/DirectionTradeSummary"
...
    DailyPerformance:
      type: object
      required: [date, trades, wins, losses, pnl, winRate]
      properties:
        date:
          type: string
        trades:
          type: integer
        wins:
          type: integer
        losses:
          type: integer
        pnl:
          type: number
        winRate:
          type: number
...
    PnlPoint:
      type: object
      required: [date, cumulativePnl, dailyPnl]
      properties:
        date:
          type: string
        cumulativePnl:
          type: number
        dailyPnl:
          type: number
...